+++ include/CeLux/backends/cpu/Decoder.hpp
@@ -8,8 +8,9 @@ namespace celux::backends::cpu
 class Decoder : public celux::Decoder
 {
   public:
-    Decoder(const std::string& filePath, std::optional<torch::Stream> stream)
-        : celux::Decoder(celux::Decoder())
+    Decoder(const std::string& filePath, std::optional<torch::Stream> stream,
+            int numThreads)
+        : celux::Decoder(stream, numThreads)
     {
         isHwAccel = false;
         initialize(filePath);