+++ include/CeLux/backends/gpu/cuda/Decoder.hpp
@@ -8,8 +8,8 @@ namespace celux::backends::gpu::cuda
 class Decoder : public celux::Decoder
 {
   public:
-    Decoder(const std::string& filePath, std::optional<torch::Stream> stream)
-        : celux::Decoder(celux::Decoder())
+    Decoder(const std::string& filePath, std::optional<torch::Stream> stream, int numThreads)
+        : celux::Decoder(stream, numThreads)
     {
         isHwAccel = true;
         initialize(filePath);